requirements_test.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements_test.txt requirements_test.in
#
attrs==21.2.0
    # via
    #   -c requirements.txt
    #   pytest
cssselect==1.1.0
    # via -r requirements_test.in
decorator==5.0.7
    # via pytest-twisted
greenlet==1.0.0
    # via pytest-twisted
iniconfig==1.1.1
    # via pytest
lxml==4.7.1
    # via -r requirements_test.in
packaging==20.9
    # via pytest
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements_test.in
    #   pytest-django
    #   pytest-twisted
pytest-django==4.5.2
    # via -r requirements_test.in
pytest-twisted==1.13.4
    # via -r requirements_test.in
toml==0.10.2
    # via pytest
